+++ packages/entities/entities-gateway-services/src/components/GatewayServiceConfigCard.vue
@@ -2,6 +2,7 @@
   <div class="kong-ui-gateway-service-entity-config-card">
     <EntityBaseConfigCard
       :config="config"
+      :config-card-doc="configCardDoc"
       :config-schema="configSchema"
       :fetch-url="fetchUrl"
       :hide-title="hideTitle"
@@ -51,7 +52,7 @@
 
       <template #tls_verify="slotProps">
         {{
-          getPropValue('rowValue', slotProps) === true || getPropValue('rowValue', slotProps) === undefined
+          typeof getPropValue('rowValue', slotProps) === 'boolean'
             ? getTlsVerifyOption('rowValue', slotProps)
             : t('gateway_services.form.fields.tls_verify_option.unset.display')
         }}
@@ -94,6 +95,14 @@ const props = defineProps({
       if (!config.entityId) return false
       return true
     },
+  },
+  /**
+   * External link for documentation that determines visibility of Documentation button
+   */
+  configCardDoc: {
+    type: String,
+    default: '',
+    required: false,
   },
   /**
    * Control visibility of card title content