+++ packages/entities/entities-routes/src/components/RouteConfigCard.vue
@@ -2,6 +2,7 @@
   <div class="kong-ui-entities-route-entity-config-card">
     <EntityBaseConfigCard
       :config="config"
+      :config-card-doc="configCardDoc"
       :config-schema="configSchema"
       :fetch-url="fetchUrl"
       :hide-title="hideTitle"
@@ -184,6 +185,14 @@ const props = defineProps({
       if (!config.entityId) return false
       return true
     },
+  },
+  /**
+   * External link for documentation that determines visibility of Documentation button
+   */
+  configCardDoc: {
+    type: String,
+    default: '',
+    required: false,
   },
   /**
    * Control visibility of card title content