+++ packages/entities/entities-gateway-services/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@1.3.2...@kong-ui-public/entities-gateway-services@1.3.3) (2023-10-16)
+
+
+### Bug Fixes
+
+* **service:** interpreting 0 as a boolean true ([#837](https://github.com/Kong/public-ui-components/issues/837)) ([3859e75](https://github.com/Kong/public-ui-components/commit/3859e758d5c5585445fdb6a0932c13f33661963b))
+
+
+
+
+
+## [1.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@1.3.1...@kong-ui-public/entities-gateway-services@1.3.2) (2023-10-13)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [1.3.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@1.3.0...@kong-ui-public/entities-gateway-services@1.3.1) (2023-10-13)
+
+
+### Bug Fixes
+
+* **service:** `tls_verify` value in config card ([#833](https://github.com/Kong/public-ui-components/issues/833)) ([96d824b](https://github.com/Kong/public-ui-components/commit/96d824bb2a29736c17dae53c011024495efa4e09))
+
+
+
+
+
+# [1.3.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@1.2.1...@kong-ui-public/entities-gateway-services@1.3.0) (2023-10-12)
+
+
+### Features
+
+* **config-card-display:** adopt in all entities [khcp-9356] ([#831](https://github.com/Kong/public-ui-components/issues/831)) ([37e8b8a](https://github.com/Kong/public-ui-components/commit/37e8b8aa00ba5eea1a113c9052ac596e1cf821fe))
+
+
+
+
+
 ## [1.2.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@1.2.0...@kong-ui-public/entities-gateway-services@1.2.1) (2023-10-10)
 
 **Note:** Version bump only for package @kong-ui-public/entities-gateway-services